Draft: Split DraftLink from Draft.py
